+++ lib/shrine/plugins/default_url.rb
@@ -6,18 +6,14 @@ class Shrine
     #
     # [doc/plugins/default_url.md]: https://github.com/shrinerb/shrine/blob/master/doc/plugins/default_url.md
     module DefaultUrl
-      def self.configure(uploader, opts = {}, &block)
-        uploader.opts[:default_url_host] = opts.fetch(:host, uploader.opts[:default_url_host])
-
-        if block
-          uploader.opts[:default_url] = block
-          Shrine.deprecation("Passing a block to default_url plugin is deprecated and will probably be removed in future versions of Shrine. Use `Attacher.default_url { ... }` instead.")
-        end
+      def self.configure(uploader, opts = {})
+        uploader.opts[:default_url] ||= {}
+        uploader.opts[:default_url].merge!(opts)
       end
 
       module AttacherClassMethods
         def default_url(&block)
-          shrine_class.opts[:default_url_block] = block
+          shrine_class.opts[:default_url][:block] = block
         end
       end
 
@@ -29,25 +25,19 @@ class Shrine
         private
 
         def default_url(**options)
-          url = if default_url_block
-                  instance_exec(options, &default_url_block)
-                elsif shrine_class.opts[:default_url]
-                  shrine_class.opts[:default_url].call(context.merge(options){|k, old, new| old})
-                end
+          return unless default_url_block
 
-          if default_url_host
-            [default_url_host, url].join
-          else
-            url
-          end
+          url = instance_exec(options, &default_url_block)
+
+          [*default_url_host, url].join
         end
 
         def default_url_block
-          shrine_class.opts[:default_url_block]
+          shrine_class.opts[:default_url][:block]
         end
 
         def default_url_host
-          shrine_class.opts[:default_url_host]
+          shrine_class.opts[:default_url][:host]
         end
       end
     end